refactor(scene): decompose scene mapper into focused modules (U2-U7)
Extract the 1,728-line desktop-agent-panel-scene.ts into cohesive
modules under scene/: tool helpers (status/title/subtitle/read-source/
result), 11 per-tool-kind payload mappers, the recursive tool-call-entry
assembler, assistant-content, conversation-model, and the chrome
builders (strips/cards/plan-sidebar/composer-model). The spine now reads
as a table of contents and re-exports the historical surface — the four
importers are untouched.

Behavior-preserving: scene characterization net (30 tests) green,
typecheck clean. GOD invariants intact (Acepe-owned display ids, no
order repair, provider types opaque).
