+++ packages/desktop/src/lib/acp/components/agent-panel/scene/tool/payloads/execute-command.ts
@@ -1,0 +1,84 @@
+import { bashHighlighter } from "../../../../../utils/bash-highlighter.svelte.js";
+
+export function splitExecuteCommandSegments(command: string): string[] {
+	const segments: string[] = [];
+	let current = "";
+	let inSingle = false;
+	let inDouble = false;
+	let index = 0;
+
+	while (index < command.length) {
+		const character = command[index];
+
+		if (character === "\\" && index + 1 < command.length) {
+			current += character + command[index + 1];
+			index += 2;
+			continue;
+		}
+
+		if (character === "'" && !inDouble) {
+			inSingle = !inSingle;
+			current += character;
+			index += 1;
+			continue;
+		}
+
+		if (character === '"' && !inSingle) {
+			inDouble = !inDouble;
+			current += character;
+			index += 1;
+			continue;
+		}
+
+		if (!inSingle && !inDouble) {
+			if (character === "&" && index + 1 < command.length && command[index + 1] === "&") {
+				const trimmed = current.trim();
+				if (trimmed) segments.push(trimmed);
+				current = "";
+				index += 2;
+				continue;
+			}
+
+			if (character === "|" && index + 1 < command.length && command[index + 1] === "|") {
+				const trimmed = current.trim();
+				if (trimmed) segments.push(trimmed);
+				current = "";
+				index += 2;
+				continue;
+			}
+
+			if (character === "|") {
+				const trimmed = current.trim();
+				if (trimmed) segments.push(trimmed);
+				current = "";
+				index += 1;
+				continue;
+			}
+
+			if (character === ";" && !(index + 1 < command.length && command[index + 1] === ";")) {
+				const trimmed = current.trim();
+				if (trimmed) segments.push(trimmed);
+				current = "";
+				index += 1;
+				continue;
+			}
+		}
+
+		current += character;
+		index += 1;
+	}
+
+	const trimmed = current.trim();
+	if (trimmed) segments.push(trimmed);
+	return segments;
+}
+
+export function mapExecuteCommandHtmls(command: string | null | undefined): readonly string[] | undefined {
+	if (!command) return undefined;
+	const segments = splitExecuteCommandSegments(command);
+	if (!bashHighlighter.ready) return undefined;
+	const highlightedSegments = segments
+		.map((segment) => bashHighlighter.highlight(segment))
+		.filter((html): html is string => html !== null);
+	return highlightedSegments.length > 0 ? highlightedSegments : undefined;
+}
