+++ packages/desktop/src/lib/acp/components/agent-panel/scene/tool/payloads/todos-payload.ts
@@ -1,0 +1,12 @@
+import type { ToolCall } from "../../../../../types/tool-call.js";
+
+export function mapTodos(toolCall: ToolCall) {
+	return toolCall.normalizedTodos?.map((todo) => {
+		return {
+			content: todo.content,
+			activeForm: todo.activeForm,
+			status: todo.status,
+			duration: todo.duration ?? null,
+		};
+	});
+}
